+++ source/mclib/inifile.h
@@ -57,7 +57,7 @@
 struct IniBlockNode
 {
 	char blockId[50];
-	unsigned long blockOffset;
+	size_t blockOffset;
 };
 
 //---------------------------------------------------------------------------
@@ -67,11 +67,11 @@ class FitIniFile : public File
 	// Data Members
 	//--------------
 	protected:
-		unsigned long 	totalBlocks;					//Total number of blocks in file
+		size_t 	totalBlocks;					//Total number of blocks in file
 		IniBlockNode 	*fileBlocks;					//Data for blocks to speed up file
 		char 				*currentBlockId;				//Id of current block
-		unsigned long 	currentBlockOffset;			//Offset into file of block start
-		unsigned long 	currentBlockSize;				//Length of current block
+		size_t 	currentBlockOffset;			//Offset into file of block start
+		size_t 	currentBlockSize;				//Length of current block
 
 	// Member Functions
 	//------------------
@@ -79,16 +79,16 @@ class FitIniFile : public File
 		long afterOpen (void);
 		void atClose (void);
 		
-		long findNextBlockStart (char *line = NULL, unsigned long lineLen = 0);
+		long findNextBlockStart (char *line = NULL, size_t lineLen = 0);
 		long countBlocks (void);
 		
-		long getNextWord (char *&line, char *buffer, unsigned long bufLen);
+		long getNextWord (char *&line, char *buffer, size_t bufLen);
 
 		float textToFloat (const char *num);
 		double textToDouble (const char *num);
 		
 		long textToLong (const char *num);
-		unsigned long textToULong (const char *num);
+		size_t textToULong (const char *num);
 		
 		short textToShort (const char *num);
 		unsigned short textToUShort (const char *num);
@@ -101,7 +101,7 @@ class FitIniFile : public File
 		float mathToFloat (const char *num);
 		
 		long mathToLong (const char *num);
-		unsigned long mathToULong (const char *num);
+		size_t mathToULong (const char *num);
 		
 		short mathToShort (const char *num);
 		unsigned short mathToUShort (const char *num);
@@ -109,25 +109,25 @@ class FitIniFile : public File
 		char mathToChar (const char *num);
 		unsigned char mathToUChar (const char *num);
 
-		long floatToText (char *result, float num, unsigned long bufLen);
+		long floatToText (char *result, float num, size_t bufLen);
 		
-		long longToTextDec (char *result, long num, unsigned long bufLen);
-		long longToTextHex (char *result, long num, unsigned long bufLen);
+		long longToTextDec (char *result, long num, size_t bufLen);
+		long longToTextHex (char *result, long num, size_t bufLen);
 
-		long shortToTextDec (char *result, short num, unsigned long bufLen);
-		long shortToTextHex (char *result, short num, unsigned long bufLen);
+		long shortToTextDec (char *result, short num, size_t bufLen);
+		long shortToTextHex (char *result, short num, size_t bufLen);
 
-		long byteToTextDec (char *result, byte num, unsigned long bufLen);	
-		long byteToTextHex (char *result, byte num, unsigned long bufLen);
+		long byteToTextDec (char *result, byte num, size_t bufLen);	
+		long byteToTextHex (char *result, byte num, size_t bufLen);
 		
-		long copyString (char* dest, char *src, unsigned long bufLen);
+		long copyString (char* dest, char *src, size_t bufLen);
 
 	public:
 		FitIniFile (void);
 		~FitIniFile (void);
 
 		virtual long open (const char* fName, FileMode _mode = READ, long numChildren = 50);
-		virtual long open (FilePtr _parent, unsigned long fileSize, long numChildren = 50);
+		virtual long open (FilePtr _parent, size_t fileSize, long numChildren = 50);
 		
 		virtual long create (char* fName);
 		virtual long createWithCase( char* fName );
@@ -147,7 +147,7 @@ class FitIniFile : public File
 		
 		long readIdBoolean (const char *varName, bool &value);
 		long readIdLong (const char *varName, long &value);
-		long readIdULong (const char *varName, unsigned long &value);
+		long readIdULong (const char *varName, size_t &value);
 		
 		long readIdShort (const char *varName, short &value);
 		long readIdUShort (const char *varName, unsigned short &value);
@@ -155,31 +155,31 @@ class FitIniFile : public File
 		long readIdChar (const char *varName, char &value);
 		long readIdUChar (const char *varName, unsigned char &value);
 		
-		long readIdString (const char *varName, char *result, unsigned long bufferSize);
+		long readIdString (const char *varName, char *result, size_t bufferSize);
 
 		long getIdStringLength (const char *varName);
 		
-		long readIdFloatArray (const char *varName, float *result, unsigned long numElements);
+		long readIdFloatArray (const char *varName, float *result, size_t numElements);
 		
-		long readIdLongArray (const char *varName, long *result, unsigned long numElements);
-		long readIdULongArray (const char *varName, unsigned long *result, unsigned long numElements);
+		long readIdLongArray (const char *varName, long *result, size_t numElements);
+		long readIdULongArray (const char *varName, size_t *result, size_t numElements);
 		
-		long readIdShortArray (const char *varName, short *result, unsigned long numElements);
-		long readIdUShortArray (const char *varName, unsigned short *result, unsigned long numElements);
+		long readIdShortArray (const char *varName, short *result, size_t numElements);
+		long readIdUShortArray (const char *varName, unsigned short *result, size_t numElements);
 		
-		long readIdCharArray (const char *varName, char *result, unsigned long numElements);
-		long readIdUCharArray (const char *varName, unsigned char *result, unsigned long numElements);
+		long readIdCharArray (const char *varName, char *result, size_t numElements);
+		long readIdUCharArray (const char *varName, unsigned char *result, size_t numElements);
 		
-		unsigned long getIdFloatArrayElements (const char *varName);
+		size_t getIdFloatArrayElements (const char *varName);
 		
-		unsigned long getIdLongArrayElements (const char *varName);
-		unsigned long getIdULongArrayElements (const char *varName);
+		size_t getIdLongArrayElements (const char *varName);
+		size_t getIdULongArrayElements (const char *varName);
 											 
-		unsigned long getIdShortArrayElements (const char *varName);
-		unsigned long getIdUShortArrayElements (const char *varName);
+		size_t getIdShortArrayElements (const char *varName);
+		size_t getIdUShortArrayElements (const char *varName);
 		
-		unsigned long getIdCharArrayElements (const char *varName);
-		unsigned long getIdUCharArrayElements (const char *varName);
+		size_t getIdCharArrayElements (const char *varName);
+		size_t getIdUCharArrayElements (const char *varName);
 		
 		long writeBlock (const char *blockId);
 
@@ -187,7 +187,7 @@ class FitIniFile : public File
 		
 		long writeIdBoolean (const char *varName, bool value);
 		long writeIdLong (const char *varName, long value);
-		long writeIdULong (const char *varName, unsigned long value);
+		long writeIdULong (const char *varName, size_t value);
 		
 		long writeIdShort (const char *varName, short value);
 		long writeIdUShort (const char *varName, unsigned short value);
@@ -197,10 +197,10 @@ class FitIniFile : public File
 		
 		long writeIdString (const char *varName, const char *result);
 
-		long writeIdFloatArray (const char *varName, float *array, unsigned long numElements);
-		long writeIdLongArray (const char *varName, long *array, unsigned long numElements);
-		long writeIdUShortArray (const char *varName, unsigned short *array, unsigned long numElements);
-		long writeIdUCharArray (const char *varName, unsigned char *array, unsigned long numElements);
+		long writeIdFloatArray (const char *varName, float *array, size_t numElements);
+		long writeIdLongArray (const char *varName, long *array, size_t numElements);
+		long writeIdUShortArray (const char *varName, unsigned short *array, size_t numElements);
+		long writeIdUCharArray (const char *varName, unsigned char *array, size_t numElements);
 };
 
 //---------------------------------------------------------------------------
